requirements.txt
pyinstaller
pycryptodome
python-dateutil